+++ onlinejudge_verify/languages/Other.py
@@ -1,0 +1,48 @@
+# Python Version: 3.x
+import pathlib
+import shlex
+import subprocess
+from logging import getLogger
+from typing import *
+
+from onlinejudge_verify.languages.base import Language
+
+logger = getLogger(__name__)
+
+class OtherLanguage(Language):
+    config: Dict[str, str]
+
+    def __init__(self, *, config: Dict[str, str]):
+        self.config = config
+
+    def compile(self, path: pathlib.Path, *, basedir: pathlib.Path, tempdir: pathlib.Path) -> None:
+        command = self.config['compile'].format(path=str(path), basedir=str(basedir), tempdir=str(tempdir))
+        logger.info('$ %s', command)
+        subprocess.check_call(shlex.split(command))
+
+    def get_execute_command(self, path: pathlib.Path, *, basedir: pathlib.Path, tempdir: pathlib.Path) -> List[str]:
+        command = self.config['execute'].format(path=str(path), basedir=str(basedir), tempdir=str(tempdir))
+        return shlex.split(command)
+
+    def list_attributes(self, path: pathlib.Path, *, basedir: pathlib.Path) -> Dict[str, str]:
+        command = self.config['list_attributes'].format(path=str(path), basedir=str(basedir))
+        text = subprocess.check_output(shlex.split(command))
+        attributes = {}
+        for line in text.splitlines():
+            key, _, value = line.decode().partition(' ')
+            attributes[key] = value
+        return attributes
+
+    def list_dependencies(self, path: pathlib.Path, *, basedir: pathlib.Path) -> List[pathlib.Path]:
+        command = self.config['list_dependencies'].format(path=str(path), basedir=str(basedir))
+        text = subprocess.check_output(shlex.split(command))
+        dependencies = [path]
+        for line in text.splitlines():
+            dependencies.append(pathlib.Path(line.decode()))
+        return dependencies
+
+    def bundle(self, path: pathlib.Path, *, basedir: pathlib.Path) -> bytes:
+        command = self.config['bundle'].format(path=str(path), basedir=str(basedir))
+        logger.info('$ %s', command)
+        return subprocess.check_output(shlex.split(command))
+
