+++ onlinejudge_verify/languages/base.py
@@ -1,0 +1,25 @@
+# Python Version: 3.x
+import abc
+import pathlib
+from typing import *
+
+class Language(object):
+    @abc.abstractmethod
+    def compile(self, path: pathlib.Path, *, basedir: pathlib.Path, tempdir: pathlib.Path) -> None:
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def get_execute_command(self, path: pathlib.Path, *, basedir: pathlib.Path, tempdir: pathlib.Path) -> List[str]:
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def list_attributes(self, path: pathlib.Path, *, basedir: pathlib.Path) -> Dict[str, str]:
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def list_dependencies(self, path: pathlib.Path, *, basedir: pathlib.Path) -> List[pathlib.Path]:
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def bundle(self, path: pathlib.Path, *, basedir: pathlib.Path) -> bytes:
+        raise NotImplementedError
